Python Mad Libs Generator
The objective of this project is to develop a Mad Libs Generator python project. In this project,
the user first has to pick a story by the title of the story. Then the user has to enter specific words like a noun,
adverb, verb, food, adjective, etc, according to the requirements. And then the story will be generated.

This is a python project for absolute beginners and is developed using the basic concept of python and tkinter.